full_backend/requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
pymongo==4.6.0
motor==3.3.2  # Async MongoDB driver

# Web scraping (existing dependencies)
requests==2.31.0
lxml==4.9.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data validation & serialization
email-validator==2.1.0

# Rate limiting & caching
slowapi==0.1.9
redis==5.0.1

# Logging & monitoring
loguru==0.7.2

# Task queue for background scraping
celery==5.3.4
redis==5.0.1

# CORS
fastapi-cors==0.0.6

# API documentation
fastapi-pagination==0.12.13

# Date handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production server
uvloop==0.19.0
httptools==0.6.1
